requirements.txt
fastapi
uvicorn
python-docx
scikit-learn
numpy
